fix(content): address review feedback (#61)
- Remove identity field setters from UserXp (keycloakUserId, productCode)
- Extract XpUpdateDto and StreakUpdateDto to shared top-level records
- Add @Nonnull to locale param in getExamBreakdown endpoint
